created student and studentservice classes based on pseudocode we created
created the student class with public properties so blazor can bind the and access the data. inside of studentservice created CRUD operations based on our pseudocode. not sure if they work, havent tested it yet. Still need to impliment UI features and remove the default files that come with creating a blazor app